+++ CHANGELOG.md
@@ -2,6 +2,21 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2020-10-04
+
+### Added
+
+Ability to use Truemail healthcheck endpoint
+
+- Added `Truemail::Client.server_healthy?`
+
+### Changed
+
+- Updated `Truemail::Client::Http#uri`
+- Updated gem dependencies
+- Updated gem documentation
+- Updated gem version
+
 ## [0.1.1] - 2020-07-21
 
 ### Changed